-refactored navbar and moved it to pages folder
