+++ components/component-index.js
@@ -1,5 +1,4 @@
 import Header from './general/header'
-import NavBar from './general/navbar'
 import StartButton from './home/start-button'
 import ExerciseScreen from './do-workout-components/screens/do-workout-exercise-screen'
 import RestScreen from './do-workout-components/screens/do-workout-rest-screen'
@@ -8,7 +7,6 @@ import CardButton from './general/cardButton'
 
 export {
     Header,
-    NavBar,
     StartButton,
     ExerciseScreen,
     RestScreen,
